Add the 'Add manually' button for resources (#359)
Adds a "Add manually" button which creates a form with required fields
for a new resource.

The user can switch between "Add manually" form and viewing the details
of an already uploaded and selected resource on the right side of the
page.

Currently it doesn't actually add the new resource. This functionality
and tests will be added in the next PR, as I also want to try some UI
changes for this page in general.