Support base64 hash checking (#472)
Added the `jsonl` output to support the Kaggle base64 example even
though it can only be run with the correct environment variables
configured. I have that setup locally and the tests pass after the
change to `download.py`. If the preference is to remove the commented
out test case, I can do that. We could setup a robot account for use in
the GH actions/CI tests, but the tests would still break for developers
unless they setup their Kaggle creds locally.

Also include a drive-by fix as the logs from the failed tests (while I
was developing) printed commands before we accounted for a parameterized
`version`.

Addresses #471