Fix flaky test (#491)
The uploadCsv test was sometimes failing to verify that the Description
field was modified.

One of the possible issues was not finding the Description field to type
into and check, as it wouldn't be visible in the current view. This was
also made more likely due to warnings to migrate usages of
`st.expermental_get_query_params`
and `st.expermental_set_query_params` which are now replaced with
`st.query_params`.
In the test itself, now waits after clilcking the "Edit fields" button,
so that the necessary input fields appear, instead of waiting after
typing and modifies the element lookup.