Do not enforce descriptions as optional.
In practical cases (as observed with both Kaggle and Hugging Face), it seems descriptions are omitted (e.g., when the name of the field is explicit). This warning used to only add noise. If we want to enforce the good practice of having descriptions filled in by users, we should probably find another way.