Add a warning in case the user deviates from the standard JSON-LD `@context` (#619)
With @goeffthomas, we realized this error is common, so we look for a
way to enforce it. This is a proposal.

```
$ mlcroissant validate --jsonld https://www.kaggle.com/datasets/kaggle/meta-kaggle/croissant/download

W0327 07:30:34.631297 140699768117056 rdf.py:77] WARNING: The JSON-LD `@context` is not standard. Refer to the official @context (e.g., from the example datasets in https://github.com/mlcommons/croissant/tree/main/datasets/1.0). The different keys are: {'isEnumeration', 'wd', 'dataCollection', 'dataBiases', 'personalSensitiveInformation'}
```

The problem is that the so-called "standard datasets" in the warning
message also don't have a proper `@context`, so I sent PR
https://github.com/mlcommons/croissant/pull/621.

```
$ mlcroissant validate --jsonld https://raw.githubusercontent.com/mlcommons/croissant/main/datasets/1.0/titanic/metadata.json

W0327 07:31:57.376785 140040957675328 rdf.py:77] WARNING: The JSON-LD `@context` is not standard. Refer to the official @context (e.g., from the example datasets in https://github.com/mlcommons/croissant/tree/main/datasets/1.0). The different keys are: {'dataCollection', 'wd', 'personalSensitiveInformation', 'dataBiases'}
W0327 07:31:57.444254 140040957675328 datasets.py:32] Found the following 1 warning(s) during the validation:
  -  [Metadata(Titanic)] Property "https://schema.org/datePublished" is recommended, but does not exist.
I0327 07:31:57.444491 140040957675328 validate.py:53] Done.
```