Include isArray and arrayShape (#807)
Addresses: #649

Update: the final implementation uses:

```
{
          "@type": "cr:Field",
          "@id": "recordset/fieldname",
          "isArray": true,
          "arrayShape": "1,2,1"
       ...
}
``` 

======

Note that this is a bit different than the original proposal in #649 :

```
{
          "@type": "cr:Field",
          "@id": "recordset/fieldname",
          "isArray": true,
          "arrayShape": [
            1,
            2,
            1
          ]
       ...
}
```

This is a bit difficult for us to process using mlcroissant... When we
create the [graph from the
jsonld](https://github.com/mlcommons/croissant/blob/d98ba0e31c7d8152c46e5a6d81dc4a017d8b1741/python/mlcroissant/mlcroissant/_src/core/json_ld.py#L210)
data, we would loose the order information; moreover, as we create a
triple for each element in the `arrayShape`, and triples need to be
unique, we would loose dimensions which have the same size.

As an alternative I propose to use a `@list`, so we keep the order info
more nicely and we don't loose data:

```
{
          "@type": "cr:Field",
          "@id": "recordset/fieldname",
          "isArray": true,
          "arrayShape": {
            "@list": [
              1,
              2,
              1
            ]
          },
       ...
}
```

WDYT?


In this PR:
* Introduces `isArray` and `arrayShape`
* Includes one example dataset
(datasets/1.0/huggingface-baratilab-flow3d/metadata.json) which is a
simplified version of
https://huggingface.co/datasets/baratilab/Flow3D-V0-Ti-6Al-4V_fast_test
(less fields, only one config) which includes the new fields.
* Updates all contexts

Still TODO: Remove `repeated` from all datasets and use `isArray` and
`arrayShape` instead.