[editor] Added support for displaying and editing `RecordSet` data types and `Field` equivalentProperty (#819)
Currently those features are not observable or editable in the editor,
but they are important to give clearer meaning to the records and
fields.

- For now the RecordSet dataTypes are shown in 1 text box with each data
type separated by a comma. It could be changed to a button and multiple
fields, but that would be a lot more code without necessarily be more
readable, let us know how do you feel. Ideally there should be some kind
of autocomplete combobox that takes multiples values, with a predefined
list of types to select from, and still the possibility to add any types
the user wants. But we could not find such component for now, so we went
for the simpler solution.
- The `equivalentProperty` has been added right after the Field dataType
in the table on the left side in the Record Sets tab, as well as in the
field details on the right side. We can remove it from the left side
table if you feel it is too much and should only be shown in the details

Doing so made us realize some confusing things with the editor example,
e.g. with the titanic dataset the RecordSets `genders` and
`embarkation_ports` have the type https://schema.org/Enumeration but the
checkbox "The RecordSet is an enumeration" is not checked. Wouldn't it
make sense to be automatically checked in this case?

We did this during a short hackathon at the https://www.swat4ls.org/
conference (semantic web for life sciences).

It is a beginning of answer to this issue:
https://github.com/mlcommons/croissant/issues/739 raised by @benjelloun
 
A lot of people in our community (semantic web for life science
research) are interested by the features requested in this issue

We followed the official specs definition for `RecordSet` and `Field`:
https://docs.mlcommons.org/croissant/docs/croissant-spec.html#field

We tested the new features using the docker deployment locally and they
work as expected. RecordSet Data types and Field equivalentProperty are
properly added to the exported JSON-LD file

It was done with @vemonet 

Also note that in the code for `Field` you use camelcase
`equivalentProperty` instead of the usual snake case that is used
elsewhere, might want to fix this for overall consistency
https://github.com/mlcommons/croissant/blob/main/python/mlcroissant/mlcroissant/_src/structure_graph/nodes/field.py#L71