1.1 chore: Verify and fix link urls (#929)
#923 

I went through the draft and made sure external links worked. Everything
seems to work, with the final urls being the same most of the time.
This PR:

* Fix a URL typo
* Adjust book reference to use DOI url

Questions:

* Most places reference schemas with `http://`, but some use `https://`.
How should schema urls be referenced in paragraphs and examples? The
links work either way, when followed the http redirects to https.
* There are some 3rd party datasets that are referenced. They still
exist, but may become dead links in the future. If possible (license
permitting) should they be replicated in the repo or placed somewhere
like torrent DHT or ipfs?