name, description language-tagged support (#932)
This proof of concept allows JSON-LD language-tagged strings for `name`
and `description` fields, fixing #924

It does not add support for general JSON-LD property-based indexing such
as id-maps or type-maps, nor does it add support for multiple titles.
i.e.

```json
{
  "name": [
    {"@value": "The Queen", "@language": "en"},
    {"@value": "Die Königin", "@language": "de"}
  ]
}
```
and
```json
{
  "name": {"en": "The Queen", "de": "Die Königin"}
}
```
are supported, but
```json
{
  "name": ["Die Königin", "Ihre Majestät"]
}
```
is not.

This implementation extends (and possibly abuses?) `field.cardinality`
to add `"LANGUAGE-TAGGED"` as a new cardinality for `name` and
`description`.

In Python and in the generated JSON-LD multilingual fields are always
represented as a [language map
(dict)](https://www.w3.org/TR/json-ld11/#language-indexing) so that
users can reference the language versions by their BCP-47 key and not
have to iterate over a list comparing `"@language"` values.

Looking for comments on the approach before adding tests, updating the
spec, ttl etc.