Add support for DICOM file format (THE format for medical images in Radiology) (#942)
This is a first part of a bigger solution, that seeks to expand
Croissant to be compatible with more LifeScience file formats ([link to
issue](https://github.com/mlcommons/croissant/issues/862)).

Now specifically, for medical images from the domain of Radiology.

Part of this PR:
  •  Adds EncodingFormat.DICOM = "image/dicom".
• Adds deps.pydicom (lazy optional import) and extras: dicom =
["pydicom>=2.4.0"]; includes in dev extras.
  •  Integrates DICOM into Read:
• _read_dicom_file uses pydicom.dcmread and returns a one-row DataFrame
with FileProperty.content holding the decoded pixel ndarray.
  •  Adds two tests:
  •  Successful minimal DICOM read.
  •  Helpful ImportError when pydicom is missing.
  
Exemplary Croissant file is here (it pulls from Zenodo):
[mri_head.json](https://github.com/user-attachments/files/22099254/mri_head.json)