Adding WildChat-1M, sQuad V2 and, Data Provenance initaitive dataset examples implementing the provenance mechanism. (#970)
**Summary**

Beyond the guidelines provided in the upcoming specification for the
provenance mechanism, it would be helpful to include examples of popular
datasets that implement this mechanism.

These examples will simplify the process of adapting the mlcroissant
library and assist adopters in integrating this feature effectively.

Changes in this PR

**WildChat-1M** dataset: Enhanced with provenance information.

**sQuad V2** dataset: Added a minimal example demonstrating the linkage
to the previous version of the sQuad dataset.

**[Common Pile: Data Provenance
Initiative:](https://huggingface.co/datasets/common-pile/data_provenance_initiative_filtered)**:
Implementing the provenance data-level mechanism.

These additions provide practical references for implementing the
provenance mechanism in other datasets.